cleaned out katwikOpsys.h + minor changes
