+++ .gitignore
@@ -5,5 +5,5 @@
 # bin
 /obj
 
-# vs:
+# vs
 /.vscode
