+++ README.md
@@ -1,6 +1,5 @@
 # TODO:
 - remove ancientCrap.h
-- remove easyAttr.h
-- add 'randAndSleep.h' and 'randAndSleep.c', clean out katwikOpsys.h
 - remove `HEADER_OKAY`
 - remove list's `*tail`
+- add `make zip` and `make submit`
