+++ littleHelpers.h
@@ -1,30 +1,15 @@
 #ifndef LITTLE_HELPERS_H
 #define LITTLE_HELPERS_H
 
-#include <stdlib.h>
+#include <sys/wait.h>
 
-// dodges ASAN sometimes :)
-// a similar macro should be in safe-malloc.h but I can't include that for some reason :(
-#define FREE(ptr) do {\
-	free(ptr);\
-	(ptr) = NULL;\
-} while(0)
+#include "easyCheck.h"
 
-// big number
-#define GIGA 1000000000L
+extern void waitAllChildren(void);
 
-// long story short, block comments don't nest nicely lol
-#define TEMP_REMOVE 0
+extern void printBuf(char* buffer, size_t sz);
 
-// useful to force stringify some stuff
-// 2 layers because uhhhh umm
-// something something stackoverflow
-#define TO_STR_(s) #s
-#define TO_STR(X) TO_STR_(X)
-	
-// goodbye Wunused-parameter/Wunused-variable lol
-#define UNUSED(thing) do {\
-	(void) (thing);\
-} while(0)
+// easy way to set a signal handler AND error check
+extern void sethandler(void (*signalHandler) (int), int sigNo);
 
 #endif // LITTLE_HELPERS_H
