+++ src/stringCrap.c
@@ -99,7 +99,7 @@ char* MY_NON_NULL(1)
 	}
 
 char* MY_NON_NULL(1)
-	strReverse(char* str, size_t length) {
+	reverse(char* str, size_t length) {
 		char* ret = str;
 
 		for (char *end = str + length - 1, tmp; str < end; ++str, --end) {
